svl/requirements.txt
python-telegram-bot==13.11.0
pymongo==4.1.1
dnspython==2.2.1
requests==2.27.1